project/requirements.txt
Django
psycopg2-binary
celery[redis]
rpi_lgpio
django-bootstrap5
